US8659: refactoring - move family stuff out of signin service, controller to family service, controller
