+++ SignInCheckIn/SignInCheckIn/Controllers/ChildSigninController.cs
@@ -21,21 +21,17 @@ namespace SignInCheckIn.Controllers
     {
         private readonly IWebsocketService _websocketService;
         private readonly IChildSigninService _childSigninService;
-        private readonly IChildCheckinService _childCheckinService;
         private readonly IKioskRepository _kioskRepository;
         private readonly IContactRepository _contactRepository;
-        private readonly IParticipantRepository _participantRepository;
-        private readonly IApplicationConfiguration _applicationConfiguration;
+        private readonly IFamilyService _familyService;
 
-        public ChildSigninController(IChildSigninService childSigninService, IWebsocketService websocketService, IChildCheckinService childCheckinService, IAuthenticationRepository authenticationRepository, IKioskRepository kioskRepository, IContactRepository contactRepository, IParticipantRepository participantRepository, IApplicationConfiguration applicationConfiguration) : base(authenticationRepository)
+        public ChildSigninController(IChildSigninService childSigninService, IWebsocketService websocketService, IAuthenticationRepository authenticationRepository, IKioskRepository kioskRepository, IContactRepository contactRepository, IFamilyService familyService) : base(authenticationRepository)
         {
             _websocketService = websocketService;
             _childSigninService = childSigninService;
-            _childCheckinService = childCheckinService;
             _kioskRepository = kioskRepository;
             _contactRepository = contactRepository;
-            _participantRepository = participantRepository;
-            _applicationConfiguration = applicationConfiguration;
+            _familyService = familyService;
         }
 
         [HttpGet]
@@ -224,47 +220,6 @@ namespace SignInCheckIn.Controllers
 
         [HttpPost]
         [ResponseType(typeof(int))]
-        [VersionedRoute(template: "signin/newfamily", minimumVersion: "1.0.0")]
-        [Route("signin/newfamily")]
-        public IHttpActionResult CreateNewFamily(List<NewParentDto> newParents)
-        {
-            return Authorized(token =>
-            {
-                string kioskIdentifier;
-
-                // make sure kiosk is admin type and configured for printing
-                if (Request.Headers.Contains("Crds-Kiosk-Identifier"))
-                {
-                    kioskIdentifier = Request.Headers.GetValues("Crds-Kiosk-Identifier").First();
-                    var kioskConfig = _kioskRepository.GetMpKioskConfigByIdentifier(Guid.Parse(kioskIdentifier));
-                    // must be kiosk type admin and have a printer set up
-                    if (kioskConfig.PrinterMapId == null || kioskConfig.KioskTypeId != 3)
-                    {
-                        throw new HttpResponseException(System.Net.HttpStatusCode.PreconditionFailed);
-                    }
-                }
-                else
-                {
-                    throw new HttpResponseException(System.Net.HttpStatusCode.PreconditionFailed);
-                }
-
-                try
-                {
-                    var participants = _childSigninService.CreateNewFamily(token, newParents, kioskIdentifier);
-                    //TODO: Figure out if this still needs to be in here for the websockets stuff
-                    //PublishSignedInParticipantsToRooms(participants);
-                    return Ok(participants);
-                }
-                catch (Exception e)
-                {
-                    var apiError = new ApiErrorDto("Create new family error: ", e);
-                    throw new HttpResponseException(apiError.HttpResponseMessage);
-                }
-            });
-        }
-
-        [HttpPost]
-        [ResponseType(typeof(int))]
         [VersionedRoute(template: "signin/family/{householdid}/member", minimumVersion: "1.0.0")]
         [Route("signin/family/{householdid}/member")]
         public IHttpActionResult AddNewFamilyMember([FromUri(Name = "householdid")] int householdId, [FromBody] ContactDto newFamilyContact)
@@ -296,7 +251,7 @@ namespace SignInCheckIn.Controllers
                         newFamilyContact
                     };
 
-                    var newParticipants = _childSigninService.AddFamilyMembers(token, householdId, newContacts);
+                    var newParticipants = _familyService.AddFamilyMembers(token, householdId, newContacts);
                     _childSigninService.CreateGroupParticipants(token, newParticipants);
                     return Ok();
                 }
